+++ lib/tar.mli
@@ -66,7 +66,16 @@ module Header : sig
     }
     (** Represents a "Pax" extended header. *)
 
-    val unmarshal : Cstruct.t -> t
+    (** [make ()] creates an extended header. *)
+    val make : ?access_time:int64 -> ?charset:string -> ?comment:string -> ?group_id:int -> ?gname:string -> ?header_charset:string -> ?link_path:string -> ?mod_time:int64 -> ?path:string -> ?file_size:int64 -> ?user_id:int -> ?uname:string -> unit -> t
+
+    (** Pretty-print the header record. *)
+    val to_detailed_string : t -> string
+
+    (** Unmarshal a pax Extended Header block. This header block may
+        be preceded by [?global] blocks which will override some
+        fields. *)
+    val unmarshal : ?global:t -> Cstruct.t -> t
   end
 
   (** Represents a standard archive (note checksum not stored). *)
@@ -92,7 +101,7 @@ module Header : sig
       [file_mode] defaults to [0o400], [user_id], [group_id] default to [0],
       [mod_time] defaults to [0L] (epoch), [link_indicator] defaults to [Normal],
       [link_name], [uname] and [gname] default to [""], and [devmajor] and
-      [devminor] default to [0].*)
+      [devminor] default to [0]. *)
   val make : ?file_mode:int -> ?user_id:int -> ?group_id:int -> ?mod_time:int64 -> ?link_indicator:Link.t -> ?link_name:string -> ?uname:string -> ?gname:string -> ?devmajor:int -> ?devminor:int -> string -> int64 -> t
 
   (** Length of a header block. *)
@@ -156,12 +165,14 @@ module HeaderReader(Async: ASYNC)(Reader: READER with type 'a t = 'a Async.t) :
   (** Returns the next header block or throws {!Header.End_of_stream} if two consecutive
       zero-filled blocks are discovered. Assumes stream is positioned at the
       possible start of a header block.
+      @param global Holds the current global pax extended header, if
+        any. Needs to be given to the next call to [read].
       @raise Header.End_of_stream if the stream unexpectedly fails. *)
-  val read : ?level:Header.compatibility -> Reader.in_channel -> (Header.t, [`Eof]) result Async.t
+  val read : ?level:Header.compatibility -> ?global:Header.Extended.t -> Reader.in_channel -> (Header.t * Header.Extended.t option, [`Eof]) result Async.t
 end
 
 module HeaderWriter(Async: ASYNC)(Writer: WRITER with type 'a t = 'a Async.t) : sig
-  val write : ?level:Header.compatibility -> Header.t -> Writer.out_channel -> unit Async.t
+  val write : ?level:Header.compatibility -> ?global:Header.Extended.t -> Header.t -> Writer.out_channel -> unit Async.t
 end
 
 module type IO = sig
@@ -187,7 +198,7 @@ module Make (IO : IO) : sig
       zero-filled blocks are discovered. Assumes stream is positioned at the
       possible start of a header block.
       @raise Stdlib.End_of_file if the stream unexpectedly fails. *)
-  val get_next_header : ?level:Header.compatibility -> IO.in_channel -> Header.t
+  val get_next_header : ?level:Header.compatibility -> ?global:Header.Extended.t -> IO.in_channel -> Header.t * Header.Extended.t option
 
   val write_block: ?level:Header.compatibility -> Header.t -> (IO.out_channel -> unit) -> IO.out_channel -> unit
     [@@ocaml.deprecated "Deprecated: use Tar.HeaderWriter"]
@@ -201,10 +212,12 @@ module Make (IO : IO) : sig
   module Archive : sig
     (** Utility functions for operating over whole tar archives. *)
 
-    (** Read the next header, apply the function 'f' to the fd and the header. The function
-        should leave the fd positioned immediately after the datablock. Finally the function
-        skips past the zero padding to the next header. *)
-    val with_next_file : IO.in_channel -> (IO.in_channel -> Header.t -> 'a) -> 'a
+    (** Read the next header, apply the function [f] to the fd, the global pax extended
+        header (if any), and the current header. The function should leave the fd positioned
+        immediately after the datablock. Finally the function skips past the zero padding to
+        the next header *)
+    val with_next_file : IO.in_channel -> ?global:Header.Extended.t ->
+                         (IO.in_channel -> Header.Extended.t option -> Header.t -> 'a) -> 'a
 
     (** List the contents of a tar. *)
     val list : ?level:Header.compatibility -> IO.in_channel -> Header.t list