+++ unix/tar_unix.ml
@@ -92,16 +92,16 @@ module Archive = struct
     extract_gen dest ifd
 
   let transform ?level f ifd ofd =
-    let rec loop () =
-      match get_next_header ifd with
+    let rec loop global () =
+      match get_next_header ?global ifd with
       | exception Tar.Header.End_of_stream -> ()
-      | header' ->
+      | (header', global) ->
         let header = f header' in
         let body = fun _ -> copy_n ifd ofd header.Tar.Header.file_size in
         write_block ?level header body ofd;
         skip ifd (Tar.Header.compute_zero_padding_length header');
-        loop () in
-    loop ();
+        loop global () in
+    loop None ();
     write_end ofd
 
   (** Create a tar on file descriptor fd from the filename list