+++ unix/tar_unix.mli
@@ -28,7 +28,7 @@ val really_write: Unix.file_descr -> Cstruct.t -> unit
     zero-filled blocks are discovered. Assumes stream is positioned at the
     possible start of a header block.
     @raise Stdlib.End_of_file if the stream unexpectedly fails. *)
-val get_next_header : ?level:Tar.Header.compatibility -> Unix.file_descr -> Tar.Header.t
+val get_next_header : ?level:Tar.Header.compatibility -> ?global:Tar.Header.Extended.t -> Unix.file_descr -> Tar.Header.t * Tar.Header.Extended.t option
 
 (** Return the header needed for a particular file on disk. *)
 val header_of_file : ?level:Tar.Header.compatibility -> string -> Tar.Header.t
@@ -48,7 +48,7 @@ module Archive : sig
   (** Read the next header, apply the function 'f' to the fd and the header. The function
       should leave the fd positioned immediately after the datablock. Finally the function
       skips past the zero padding to the next header. *)
-  val with_next_file : Unix.file_descr -> (Unix.file_descr -> Tar.Header.t -> 'a) -> 'a
+  val with_next_file : Unix.file_descr -> ?global: Tar.Header.Extended.t -> (Unix.file_descr -> Tar.Header.Extended.t option -> Tar.Header.t -> 'a) -> 'a
 
   (** List the contents of a tar. *)
   val list : ?level:Tar.Header.compatibility -> Unix.file_descr -> Tar.Header.t list